requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
google-api-python-client==2.88.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
sqlalchemy==2.0.23
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0